[Feat] - 온보딩 전체적인 플로우 및 등록하기 버튼으로 탭바로 이동 구현 완료
- 기타 리팩토링
